[REFACTOR] Move add challenge flow into new component
