+++ src/client/react/admin/views/ChallengesView/index.js
@@ -3,15 +3,14 @@ import PropTypes from 'prop-types';
 import autobind from 'core-decorators/es/autobind';
 import { connect } from 'react-redux';
 import { graphql, compose } from 'react-apollo';
-import { Spinner, Button, Intent, Dialog } from '@blueprintjs/core';
-import { getAllChallenges, createChallenge } from '../../../../graphql/challenge';
+import { Spinner, Intent } from '@blueprintjs/core';
+import { getAllChallenges } from '../../../../graphql/challenge';
 import { saveState } from '../../../../actions/stateActions';
 import RefreshBar from '../../components/RefreshBar';
 import ViewError from '../../components/ViewError';
 import ChallengeCard from './ChallengeCard';
 import ChallengeProfile from './ChallengeProfile';
-import NotificationToaster from '../../../components/NotificationToaster';
-import FormInput from '../../../../../../lib/react/components/forms/FormInput';
+import AddChallenge from './AddChallenge';
 
 import '../../scss/views/_challenges-view.scss';
 
@@ -21,10 +20,7 @@ const QueryGetAllChallengesOptions = {
 	options: { networkPolicy: 'cache-and-network' }
 }
 
-@compose(
-	graphql(getAllChallenges('_id key order title public locked items{key}'), QueryGetAllChallengesOptions),
-	graphql(createChallenge('ok'), { name: 'MutationCreateChallenge' })
-)
+@graphql(getAllChallenges('_id key order title public locked items{key}'), QueryGetAllChallengesOptions)
 @connect()
 @autobind
 class ChallengesView extends React.Component {
@@ -34,12 +30,7 @@ class ChallengesView extends React.Component {
 
 	state = {
 		viewProfile: null,
-		refetching: false,
-		showCreateChallenge: false,
-		createChallengeLoading: false,
-		createChallengeError: null,
-		challengeKey: '',
-		order: '0'
+		refetching: false
 	}
 
 	componentDidMount() {
@@ -77,47 +68,6 @@ class ChallengesView extends React.Component {
 		});
 	}
 
-	toggleCreateChallenge() {
-		this.setState((prevState) => {
-			return { showCreateChallenge: !prevState.showCreateChallenge }
-		});
-	}
-	
-	editValue(valueName) {
-		return (e) => {
-			this.setState({ [valueName]: e.target.value });
-		}
-	}
-
-	async submitCreateChallenge() {
-		this.setState({ createChallengeLoading: true, createChallengeError: null });
-		try {
-			const id = parseInt(Date.now()/1000) % 100000;
-			await this.props.MutationCreateChallenge({
-				variables: {
-					key: this.state.challengeKey,
-					order: this.state.order,
-					passphrase: `${id}`,
-					title: `New Challenge #${id}`,
-					description: `## Challenge ${id}`
-				}
-			});
-			await this.props.QueryGetAllChallenges.refetch();
-			this.setState({ createChallengeLoading: false, showCreateChallenge: false, challengeKey: '' });
-		}
-		catch (err) {
-			if (this._mounted) {
-				this.setState({ createChallengeLoading: false, createChallengeError: err.toString() });
-			}
-			else {
-				NotificationToaster.show({
-					intent: Intent.DANGER,
-					message: err.toString()
-				});
-			}
-		}
-	}
-
 	render() {
 		let content = null;
 
@@ -135,8 +85,7 @@ class ChallengesView extends React.Component {
 					{ getAllChallenges.map((challenge) => {
 						return <ChallengeCard key={challenge.key} challenge={challenge} renderProfile={this.renderProfile}/>;
 					})}
-					<Button text='Create new challenge' iconName='map' className='pt-fill pt-minimal' 
-						intent={Intent.PRIMARY} onClick={this.toggleCreateChallenge}/>
+					<AddChallenge refetch={this.props.QueryGetAllChallenges.refetch}/>
 				</div>
 			);
 		}
@@ -150,36 +99,6 @@ class ChallengesView extends React.Component {
 				<RefreshBar query={this.props.QueryGetAllChallenges} refetching={this.state.refetching} 
 					disabled={this.state.viewProfile} shouldRefresh={this.props.shouldRefresh}/>
 				{content}
-
-				{/* Create Challenge dialog */}
-				<Dialog isOpen={this.state.showCreateChallenge} title='Create a new challenge' iconName='map' onClose={this.toggleCreateChallenge}>
-					<div className='pt-dialog-body'>
-						{this.state.createChallengeError ? 
-							<div className='pt-callout pt-intent-danger pt-icon-error' style={{marginBottom:'0.5rem'}}>
-								{this.state.createChallengeError}
-							</div>
-							:null}
-						<label className='pt-label'>
-							<div className='pt-callout pt-icon-info-sign' style={{marginBottom:'0.5rem'}}>
-								The <b>challenge key</b> should be a unique identifier for this challenge,
-								and cannot be changed once set.
-								The challenge <b>order</b> determines the order in which the challenge appears on the screen.
-							</div>
-							<div>
-								<b>Challenge key:</b> 
-								<FormInput id='challenge-key' value={this.state.challengeKey} onChange={this.editValue('challengeKey')}/>
-								<b>Order:</b> 
-								<FormInput id='challenge-order' value={this.state.order} onChange={this.editValue('order')}/>
-							</div>
-						</label>
-					</div>
-					<div className='pt-dialog-footer'>
-						<div className='pt-dialog-footer-actions'>
-							<Button onClick={this.toggleCreateChallenge} text='Cancel' className='pt-minimal' disabled={this.state.createChallengeLoading}/>
-							<Button onClick={this.submitCreateChallenge} text='Create' intent={Intent.PRIMARY} loading={this.state.createChallengeLoading}/>
-						</div>
-					</div>
-				</Dialog>
 			</div>
 		);
 	}
